requirements.txt
aniso8601==7.0.0
attrs==19.3.0
Cerberus==1.3.2
click==7.1.2
falcon==3.0.0a1
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphene-sqlalchemy-filter==1.10.2
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
importlib-metadata==1.6.0
more-itertools==8.3.0
mysqlclient==1.4.6
packaging==20.4
pluggy==0.13.1
promise==2.3
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.2
python-engineio==3.12.1
python-socketio==4.5.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.17
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
zipp==3.1.0
